requirements.txt
django==4.0
django-money==2.1
djangorestframework==3.13.0
pyyaml==6.0
uritemplate==4.1.1
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
